Extract NZBGet connection testing
